Rename loaders to sync
